Improve searchable content extraction in PageSnapshotToElasticaTransformer
The extraction method of searchable content from the layout block has been adjusted in the PageSnapshotToElasticaTransformer. Now, the code first checks if the class implements the searchableContentInterface or has a 'getSearchableContent' method before deserializing the snapshot content. Furthermore, if the content item exists in the managerRegistry, it will be refreshed to ensure up-to-dateness of the indexed content.